+++ src/components/professional-interviews/index.ts
@@ -2,3 +2,4 @@ export { ProfessionalCampaignList } from './ProfessionalCampaignList';
 export { ProfessionalTeamView } from './ProfessionalTeamView';
 export { ProfessionalInterviewView } from './ProfessionalInterviewView';
 export { MyProfessionalInterviewsView } from './MyProfessionalInterviewsView';
+export { ProfessionalEmployeeHistoryView } from './ProfessionalEmployeeHistoryView';